requirements.txt
alembic==1.2.1
androguard==3.3.5
asn1crypto==1.2.0
backcall==0.1.0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-Dropzone==1.5.4
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
flask-paginate==0.5.3
Flask-QRcode==3.0.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
kiwisolver==1.1.0
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.4
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.10
pydot==1.4.1
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.5
python-dateutil==2.8.0
python-editor==1.0.4
python-memcached==1.59
qiniu==7.2.4
qrcode==6.1
redis==3.3.8
requests==2.22.0
shortuuid==0.5.0
six==1.12.0
SQLAlchemy==1.3.9
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
